The building's architect would like to take a break...
A majority of the roomstuff is now implemented.  The minimap ACTUALLY
WORKS NOW (mostly the URHere marker and displaying numbers on the map),
and going from room to room is complete (that was an ass to code).

Fu.  When we get to generating the puzzle itself, take a look at
Assets\RoomManager.cs line 15.