The UI gets a fresh look...
Redid the game layout.  There's now only a single pane of GUI and there
are four sub-panes, one for each player.  If we feel that we can't get
in multiplayer by the end of the project, it should be easy to make it
stick to just one player.

Half-tempted to do an input helper class.  Chances are we won't want
players to change the inputs if we're doing 360 controllers.